Make model price sync source interface-driven
